chore!: Move metadata to engine data (#18)
This changes the internal storage of Metadata to use EngineData instead of converting the data to structs. It adds
a method to do the conversion so unit tests can be kept relatively consistent.

There was some potential foot-gun in the code that tried to stat the file size from object storage. I changed the write_parquet API slightly to move this as an engine detail (it seems the upstream pull request already does this.